pyproject.toml
[project]
name = "google-conversational-agents-whatsapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "fastapi>=0.119.1",
    "google-cloud-dialogflow-cx>=1.43.0",
    "google-genai>=1.46.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
